pyproject.toml
[tool.poetry]
name = "better-roadmap"
version = "0.1.0"
description = "Create a Gantt chart-y longterm roadmap for a scrum project"
authors = ["Rodrigo Eliseu <rodrigo@eliseu.me>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.0"
plotly-express = "^0.4.1"
Flask = "^2.0.2"
PyYAML = "^6.0"
dash = "^2.1.0"
gunicorn = "^20.1.0"
dash-bootstrap-components = "^1.0.3"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"
flake8-bugbear = "^22.1.11"
ipython = "^8.0.1"
pytest = "^6.2.5"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
